프로필 수정 로직 변경 (#162)
* refactor: 프로필 관련 로직 Profile 모델로 이전

* refactor: 프로필 관련 serialerzer Profile 모델로 이전

* refactor: 프로필 수정 serializer 변경

* refactor: 프로필 view 로직 Profile 모델로 이전
