+++ backend/account/serializers.py
@@ -44,14 +44,13 @@ class UserRegisterSerializer(serializers.Serializer):
 
 class UserChangePasswordSerializer(serializers.Serializer):
     old_password = serializers.CharField()
-    new_password = serializers.CharField(min_length=6)
-    tfa_code = serializers.CharField(required=False, allow_blank=True)
-
-
-class UserChangeEmailSerializer(serializers.Serializer):
-    password = serializers.CharField()
-    new_email = serializers.EmailField(max_length=64)
-    tfa_code = serializers.CharField(required=False, allow_blank=True)
+    new_password = serializers.RegexField(
+        regex=r'^(?=.*[A-Za-z])(?=.*\d)(?=.*[@$!%*#?&])[A-Za-z\d@$!%*#?&]{8,}$',
+        min_length=8,
+        error_messages={
+            'invalid': '비밀번호는 8글자 이상이어야 하며, 영문, 숫자, 특수문자를 모두 포함해야 합니다.'
+        }
+    )
 
 
 class GenerateUserSerializer(serializers.Serializer):
@@ -64,21 +63,31 @@ class GenerateUserSerializer(serializers.Serializer):
 class ImportUserSeralizer(serializers.Serializer):
     users = serializers.ListField(
         child=serializers.ListField(child=serializers.CharField(max_length=64)))
+
+
 class DateRangeSerializer(serializers.Serializer):
     start = serializers.DateField()
     end = serializers.DateField()
+
+
 class WeeklyStatisticsSerializer(serializers.Serializer):
     week_info = serializers.CharField()
     date_range = DateRangeSerializer()
     xAxis = serializers.ListField(child=serializers.CharField())
     series = serializers.ListField(child=serializers.IntegerField())
+
+
 class MonthlyStatisticsSerializer(serializers.Serializer):
     year = serializers.IntegerField()
     xAxis = serializers.ListField(child=serializers.CharField())
     series = serializers.ListField(child=serializers.IntegerField())
+
+
 class DepartmentStatSerializer(serializers.Serializer):
     value = serializers.IntegerField()
     name = serializers.CharField()
+
+
 class UserAdminStatisticsSerializer(serializers.Serializer):
     all_users = serializers.IntegerField()
     super_admins = serializers.IntegerField()
@@ -89,6 +98,7 @@ class UserAdminStatisticsSerializer(serializers.Serializer):
     monthly_statistics = MonthlyStatisticsSerializer()
     weekly_statistics = WeeklyStatisticsSerializer()
 
+
 class UserAdminSerializer(serializers.ModelSerializer):
     real_name = serializers.SerializerMethodField()
     college = serializers.SerializerMethodField()
@@ -98,7 +108,7 @@ class UserAdminSerializer(serializers.ModelSerializer):
 
     class Meta:
         model = User
-        fields = ["id", "username", "email", "admin_type", "problem_permission", "real_name","college", "department",
+        fields = ["id", "username", "email", "admin_type", "problem_permission", "real_name", "college", "department",
                   "create_time", "last_login", "two_factor_auth", "open_api", "is_disabled", "avatar", "student_id"]
 
     def get_real_name(self, obj):
@@ -116,121 +126,12 @@ class UserAdminSerializer(serializers.ModelSerializer):
     def get_student_id(self, obj):
         return obj.userprofile.student_id
 
+
 class UserSerializer(serializers.ModelSerializer):
     class Meta:
         model = User
         fields = ["id", "username", "email", "admin_type", "problem_permission",
                   "create_time", "last_login", "two_factor_auth", "open_api", "is_disabled"]
-
-
-class UserProfileSerializer(serializers.ModelSerializer):
-    user = UserSerializer()
-    real_name = serializers.SerializerMethodField()
-
-    class Meta:
-        model = UserProfile
-        fields = "__all__"
-
-    def __init__(self, *args, **kwargs):
-        self.show_real_name = kwargs.pop("show_real_name", False)
-        super(UserProfileSerializer, self).__init__(*args, **kwargs)
-
-    def get_real_name(self, obj):
-        return obj.real_name if self.show_real_name else None
-
-
-class DashboardUserInfoSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = User
-        fields = ['username', 'email']
-
-
-class DashboardSubmissionSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = UserProfile
-        fields = ['accepted_number', 'submission_number']
-
-
-class DashboardCollegeSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = College
-        fields = ['college_name']
-
-
-class DashboardDepartmentSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Department
-        fields = ['department_name']
-
-
-class DashboardRankSerializer(serializers.ModelSerializer):
-    total_rank = serializers.SerializerMethodField()
-    total_rank_percentage = serializers.SerializerMethodField()
-
-    class Meta:
-        model = UserScore
-        fields = ['tier', 'total_rank', 'total_rank_percentage', 'total_score', 'current_tier_score', 'next_tier_score']
-
-    def get_total_rank(self, instance):
-        return instance.total_rank
-
-    def get_total_rank_percentage(self, instance):
-        total_rank_percentage = round(instance.total_rank / self.context['total_user_count'], 2)
-        return total_rank_percentage
-
-
-class DashboardFieldInfoSerializer(serializers.ModelSerializer):
-    fieldInfo = serializers.SerializerMethodField()
-
-    class Meta:
-        model = UserScore
-        fields = ['fieldInfo']
-
-    def get_fieldInfo(self, instance):
-        field_info = {}
-        fields = ['datastructure', 'math', 'sorting', 'implementation', 'search']
-
-        for field in fields:
-            field_name_score = f"{field}_score"
-            user_field_score = getattr(instance, field_name_score)
-
-            field_rank = UserScore.objects.filter(**{f'{field_name_score}__gt': user_field_score}).count() + 1
-            total_users = UserScore.objects.count()
-            rank_percentage = (field_rank / total_users)
-
-            field_info[field] = {
-                'score': user_field_score,
-                'ranking': field_rank,
-                'ranking_percent': rank_percentage,
-            }
-
-        return field_info
-
-
-class DashboardDifficultyInfoSerializer(serializers.ModelSerializer):
-    difficultyInfo = serializers.SerializerMethodField()
-
-    class Meta:
-        model = UserSolved
-        fields = ['difficultyInfo']
-
-    def get_difficultyInfo(self, instance):
-        difficulty_info = {}
-        difficulties = ['VeryLow', 'Low', 'Mid', 'High', 'VeryHigh']
-
-        for difficulty in difficulties:
-            solved_field = f"{difficulty}_solved"
-            score_field = f"{difficulty}_score"
-
-            solved_count = getattr(instance, solved_field)
-            total_score = getattr(UserScore.objects.get(user=instance.user), score_field)
-
-            difficulty_info[difficulty.lower()] = {
-                'solve_number': solved_count,
-                'total_score': total_score,
-            }
-
-        return difficulty_info
 
 
 class EditUserSerializer(serializers.Serializer):
@@ -269,17 +170,6 @@ class EditUserSerializer(serializers.Serializer):
     is_disabled = serializers.BooleanField()
 
 
-class EditUserProfileSerializer(serializers.Serializer):
-    real_name = serializers.CharField(max_length=32, allow_null=True, required=False)
-    avatar = serializers.CharField(max_length=256, allow_blank=True, required=False)
-    blog = serializers.URLField(max_length=256, allow_blank=True, required=False)
-    mood = serializers.CharField(max_length=256, allow_blank=True, required=False)
-    github = serializers.URLField(max_length=256, allow_blank=True, required=False)
-    school = serializers.CharField(max_length=64, allow_blank=True, required=False)
-    major = serializers.CharField(max_length=64, allow_blank=True, required=False)
-    language = serializers.CharField(max_length=32, allow_blank=True, required=False)
-
-
 class ApplyResetPasswordSerializer(serializers.Serializer):
     email = serializers.EmailField()
     captcha = serializers.CharField()
@@ -300,10 +190,6 @@ class SSOSerializer(serializers.Serializer):
 
 class TwoFactorAuthCodeSerializer(serializers.Serializer):
     code = serializers.IntegerField()
-
-
-class ImageUploadForm(forms.Form):
-    image = forms.FileField()
 
 
 class FileUploadForm(forms.Form):
