+++ backend/account/urls/oj.py
@@ -3,11 +3,9 @@ from django.conf.urls import url
 from ..views.oj import (ApplyResetPasswordAPI, ResetPasswordAPI,
                         UserChangePasswordAPI, UserRegisterAPI,
                         UserLoginAPI, UserLogoutAPI, UsernameOrEmailCheck,
-                        AvatarUploadAPI, TwoFactorAuthAPI, UserProfileAPI,
-                        UserProfileDashBoardAPI,
-                        CheckTFARequiredAPI, SessionManagementAPI,
-                        ProfileProblemDisplayIDRefreshAPI, OpenAPIAppkeyAPI, SSOAPI,
-                        ApplyUserEmailValidCheckAPI, UserEmailValidCheckAPI, NicknameValidCheckAPI)
+                        TwoFactorAuthAPI, CheckTFARequiredAPI, SessionManagementAPI,
+                        OpenAPIAppkeyAPI, SSOAPI, ApplyUserEmailValidCheckAPI,
+                        UserEmailValidCheckAPI, NicknameValidCheckAPI)
 
 from utils.captcha.views import CaptchaAPIView
 
@@ -24,10 +22,6 @@ urlpatterns = [
     url(r"^reset_password/?$", ResetPasswordAPI.as_view(), name="reset_password_api"),
     url(r"^captcha/?$", CaptchaAPIView.as_view(), name="show_captcha"),
     url(r"^check_username_or_email", UsernameOrEmailCheck.as_view(), name="check_username_or_email"),
-    url(r"^profile/?$", UserProfileAPI.as_view(), name="user_profile_api"),
-    url(r"^profile/dashboard/?$", UserProfileDashBoardAPI.as_view(), name="user_profile_dashboard_api"),
-    url(r"^profile/fresh_display_id", ProfileProblemDisplayIDRefreshAPI.as_view(), name="display_id_fresh"),
-    url(r"^upload_avatar/?$", AvatarUploadAPI.as_view(), name="avatar_upload_api"),
     url(r"^tfa_required/?$", CheckTFARequiredAPI.as_view(), name="tfa_required_check"),
     url(r"^two_factor_auth/?$", TwoFactorAuthAPI.as_view(), name="two_factor_auth_api"),
     url(r"^sessions/?$", SessionManagementAPI.as_view(), name="session_management_api"),
