+++ backend/account/views/oj.py
@@ -31,124 +31,9 @@ from ..models import User, UserProfile, AdminType, UserScore, UserSolved
 from ..serializers import (ApplyResetPasswordSerializer, ResetPasswordSerializer,
                            UserChangePasswordSerializer, UserLoginSerializer,
                            UserRegisterSerializer, UsernameOrEmailCheckSerializer,
-                           RankInfoSerializer, UserChangeEmailSerializer, SSOSerializer,
-                           DashboardSubmissionSerializer,
-                           DashboardDepartmentSerializer, DashboardCollegeSerializer, DashboardRankSerializer,
-                           DashboardUserInfoSerializer, DashboardFieldInfoSerializer,
-                           DashboardDifficultyInfoSerializer)
-from ..serializers import (TwoFactorAuthCodeSerializer, UserProfileSerializer,
-                           EditUserProfileSerializer, ImageUploadForm)
+                           RankInfoSerializer, SSOSerializer)
+from ..serializers import TwoFactorAuthCodeSerializer
 from ..tasks import send_email_async
-
-
-class UserProfileAPI(APIView):
-    @method_decorator(ensure_csrf_cookie)
-    def get(self, request, **kwargs):
-        """
-        判断是否登录， 若登录返回用户信息
-        """
-        user = request.user
-        if not user.is_authenticated:
-            return self.success()
-        show_real_name = False
-        username = request.GET.get("username")
-        try:
-            if username:
-                user = User.objects.get(username=username, is_disabled=False)
-            else:
-                user = request.user
-                # api返回的是自己的信息，可以返real_name
-                show_real_name = True
-        except User.DoesNotExist:
-            return self.error("User does not exist")
-        return self.success(UserProfileSerializer(user.userprofile, show_real_name=show_real_name).data)
-
-    @validate_serializer(EditUserProfileSerializer)
-    @login_required
-    def put(self, request):
-        data = request.data
-        user_profile = request.user.userprofile
-        for k, v in data.items():
-            setattr(user_profile, k, v)
-        user_profile.save()
-        return self.success(UserProfileSerializer(user_profile, show_real_name=True).data)
-
-
-class UserProfileDashBoardAPI(APIView):
-    @login_required
-    def get(self, request):
-        try:
-            username = request.GET.get("username")
-            user_profile = UserProfile.objects.filter(user__username=username).first()
-            user_id = user_profile.user_id
-            user_score = UserScore.objects.filter(user_id=user_id).annotate(
-                total_rank=Count('total_score',
-                                 filter=Q(total_score__gt=F('total_score'))) + 1,
-                datastructure_rank=Count('datastructure_score',
-                                         filter=Q(datastructure_score__gt=F('datastructure_score'))) + 1,
-                implementation_rank=Count('implementation_score',
-                                          filter=Q(implementation_score__gt=F('implementation_score'))) + 1,
-                math_rank=Count('math_score',
-                                filter=Q(math_score__gt=F('math_score'))) + 1,
-                search_rank=Count('search_score',
-                                  filter=Q(search_score__gt=F('search_score'))) + 1,
-                sorting_rank=Count('sorting_score',
-                                   filter=Q(search_score__gt=F('sorting_score'))) + 1
-            ).first()
-            user_solved = UserSolved.objects.filter(user_id=user_id).first()
-        except User.DoesNotExist or UserProfile.DoesNotExist:
-            return HttpResponseNotFound('user does not exist')
-        except Department.DoesNotExist or College.DoesNotExist:
-            return HttpResponseNotFound('department or college does not exist')
-        except UserScore.DoesNotExist:
-            return HttpResponseNotFound('user_score does not exist')
-
-        total_submitted_user_count = UserScore.objects.count()
-
-        """ Build oj_status """
-        ojStatus = {}
-        ojStatus.update(DashboardSubmissionSerializer(user_profile).data)
-        ojStatus.update(DashboardRankSerializer(user_score, context={'total_user_count': total_submitted_user_count}).data)
-
-        """ Build fieldInfo """
-        fieldInfo = DashboardFieldInfoSerializer(user_score).data['fieldInfo']
-
-        """ Build difficultyInfo"""
-        difficultyInfo = DashboardDifficultyInfoSerializer(user_solved).data['difficultyInfo']
-
-        response_data = {
-            'ojStatus': ojStatus,
-            'fieldInfo': fieldInfo,
-            'difficultyInfo': difficultyInfo
-        }
-        return self.success(response_data)
-
-
-class AvatarUploadAPI(APIView):
-    request_parsers = ()
-
-    @login_required
-    def post(self, request):
-        form = ImageUploadForm(request.POST, request.FILES)
-        if form.is_valid():
-            avatar = form.cleaned_data["image"]
-        else:
-            return self.error("Invalid file content")
-        if avatar.size > 2 * 1024 * 1024:
-            return self.error("Picture is too large")
-        suffix = os.path.splitext(avatar.name)[-1].lower()
-        if suffix not in [".gif", ".jpg", ".jpeg", ".bmp", ".png"]:
-            return self.error("Unsupported file format")
-
-        name = rand_str(10) + suffix
-        with open(os.path.join(settings.AVATAR_UPLOAD_DIR, name), "wb") as img:
-            for chunk in avatar:
-                img.write(chunk)
-        user_profile = request.user.userprofile
-
-        user_profile.avatar = f"{settings.AVATAR_URI_PREFIX}/{name}"
-        user_profile.save()
-        return self.success("Succeeded")
 
 
 class TwoFactorAuthAPI(APIView):
@@ -369,14 +254,9 @@ class UserChangePasswordAPI(APIView):
         User change password api
         """
         data = request.data
-        username = request.user.username
+        username = request.user.email
         user = auth.authenticate(username=username, password=data["old_password"])
         if user:
-            if user.two_factor_auth:
-                if "tfa_code" not in data:
-                    return self.error("tfa_required")
-                if not OtpAuth(user.tfa_token).valid_totp(data["tfa_code"]):
-                    return self.error("Invalid two factor verification code")
             user.set_password(data["new_password"])
             user.save()
             return self.success("Succeeded")
